[MIST-689] address comments
