+++ src/main/java/edu/snu/mist/core/task/metrics/parameters/GlobalProcCpuUtilAlpha.java
@@ -13,15 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.mist.core.parameters;
+package edu.snu.mist.core.task.metrics.parameters;
 
 import org.apache.reef.tang.annotations.Name;
 import org.apache.reef.tang.annotations.NamedParameter;
 
 /**
- * An alpha value of num events in GroupMetric.
- * It would be used as a decaing rate in EWMA.
+ A decaying rate of EWMA for process cpu utilization in CpuUtilMetric.
  */
-@NamedParameter(doc="An alpha value of num events in GroupMetric", default_value = "0.7")
-public final class GroupNumEventAlpha implements Name<Double> {
+@NamedParameter(doc="A decaying rate for process cpu utilization in CpuUtilMetric", default_value = "0.7")
+public final class GlobalProcCpuUtilAlpha implements Name<Double> {
 }
