+++ src/main/java/edu/snu/mist/core/task/metrics/parameters/HeapMemoryUsageAlpha.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (C) 2017 Seoul National University
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package edu.snu.mist.core.task.metrics.parameters;
+
+import org.apache.reef.tang.annotations.Name;
+import org.apache.reef.tang.annotations.NamedParameter;
+
+/**
+ * A decaying rate of EWMA for the heap memory usage.
+ */
+@NamedParameter(doc="An alpha value of heap memory usage metric.", default_value = "0.7")
+public final class HeapMemoryUsageAlpha implements Name<Double> {
+}
