+++ src/test/java/edu/snu/mist/core/task/globalsched/EventNumAndWeightMetricTest.java
@@ -15,7 +15,7 @@
  */
 package edu.snu.mist.core.task.globalsched;
 
-import edu.snu.mist.core.parameters.GlobalNumEventAlpha;
+import edu.snu.mist.core.task.metrics.parameters.GlobalNumEventAlpha;
 import edu.snu.mist.core.task.MetricUtil;
 import edu.snu.mist.core.task.globalsched.metrics.EventNumAndWeightMetric;
 import org.apache.reef.tang.Configuration;
